Refactor to common class AwsChunkedInputStream
