+++ core/sdk-core/src/main/java/software/amazon/awssdk/core/internal/io/UnderlyingStreamBuffer.java
@@ -20,8 +20,8 @@ import software.amazon.awssdk.core.exception.SdkClientException;
 import software.amazon.awssdk.utils.Logger;
 
 @SdkInternalApi
-class DecodedStreamBuffer {
-    private static final Logger log = Logger.loggerFor(DecodedStreamBuffer.class);
+class UnderlyingStreamBuffer {
+    private static final Logger log = Logger.loggerFor(UnderlyingStreamBuffer.class);
 
     private byte[] bufferArray;
     private int maxBufferSize;
@@ -29,7 +29,7 @@ class DecodedStreamBuffer {
     private int pos = -1;
     private boolean bufferSizeOverflow;
 
-    DecodedStreamBuffer(int maxBufferSize) {
+    UnderlyingStreamBuffer(int maxBufferSize) {
         bufferArray = new byte[maxBufferSize];
         this.maxBufferSize = maxBufferSize;
     }
